+++ rabbitizer/rabbitizer_type_Instruction.c
@@ -289,12 +289,42 @@ static PyGetSetDef rabbitizer_type_Instruction_getsetters[] = {
     }
 
 DEF_METHOD_GET_UINT(getRaw)
+
 static PyObject *rabbitizer_type_Instruction_getImmediate(PyRabbitizerInstruction *self, UNUSED PyObject *closure) {
+    if (!RabbitizerInstruction_hasOperandAlias(&self->instr, RAB_OPERAND_cpu_immediate)) {
+        PyErr_Format(PyExc_RuntimeError, "'%s' instruction does not have an 'immediate' field", RabbitizerInstrId_getOpcodeName(self->instr.uniqueId));
+        return NULL;
+    }
+
     return PyLong_FromUnsignedLong(RAB_INSTR_GET_immediate(&self->instr));
 }
-DEF_METHOD_GET_INT(getProcessedImmediate)
-DEF_METHOD_GET_UINT(getInstrIndexAsVram)
-DEF_METHOD_GET_INT(getBranchOffset)
+
+static PyObject *rabbitizer_type_Instruction_getProcessedImmediate(PyRabbitizerInstruction *self,UNUSED PyObject *closure) {
+    if (!RabbitizerInstruction_hasOperandAlias(&self->instr, RAB_OPERAND_cpu_immediate)) {
+        PyErr_Format(PyExc_RuntimeError, "'%s' instruction does not have an 'immediate' field", RabbitizerInstrId_getOpcodeName(self->instr.uniqueId));
+        return NULL;
+    }
+
+    return PyLong_FromLong(RabbitizerInstruction_getProcessedImmediate(&self->instr)); 
+}
+
+static PyObject *rabbitizer_type_Instruction_getInstrIndexAsVram(PyRabbitizerInstruction *self, UNUSED PyObject *closure) {
+    if (!RabbitizerInstruction_hasOperandAlias(&self->instr, RAB_OPERAND_cpu_label)) {
+        PyErr_Format(PyExc_RuntimeError, "'%s' instruction does not have an 'jump label' field", RabbitizerInstrId_getOpcodeName(self->instr.uniqueId));
+        return NULL;
+    }
+
+    return PyLong_FromUnsignedLong(RabbitizerInstruction_getInstrIndexAsVram(&self->instr));
+}
+
+static PyObject *rabbitizer_type_Instruction_getBranchOffset(PyRabbitizerInstruction *self, UNUSED PyObject *closure) {
+    if (!RabbitizerInstruction_hasOperand(&self->instr, RAB_OPERAND_cpu_branch_target_label)) {
+        PyErr_Format(PyExc_RuntimeError, "'%s' instruction does not have an 'branch label' field", RabbitizerInstrId_getOpcodeName(self->instr.uniqueId));
+        return NULL;
+    }
+
+    return PyLong_FromLong(RabbitizerInstruction_getBranchOffset(&self->instr));
+}
 
 static PyObject *rabbitizer_type_Instruction_getOpcodeName(PyRabbitizerInstruction *self, UNUSED PyObject *closure) {
     return PyUnicode_FromString(RabbitizerInstrId_getOpcodeName(self->instr.uniqueId));
@@ -311,8 +341,23 @@ static PyObject *rabbitizer_type_Instruction_getGenericBranchOffset(PyRabbitizer
     return PyLong_FromLong(RabbitizerInstruction_getGenericBranchOffset(&self->instr, currentVram));
 }
 
-DEF_METHOD_GET_INT(getBranchOffsetGeneric)
-DEF_METHOD_GET_UINT(getBranchVramGeneric)
+static PyObject *rabbitizer_type_Instruction_getBranchOffsetGeneric(PyRabbitizerInstruction *self, UNUSED PyObject *closure) {
+    if (!RabbitizerInstruction_hasOperandAlias(&self->instr, RAB_OPERAND_cpu_label) && !RabbitizerInstruction_hasOperand(&self->instr, RAB_OPERAND_cpu_branch_target_label)) {
+        PyErr_Format(PyExc_RuntimeError, "'%s' instruction does not not perform a branch or a jump.", RabbitizerInstrId_getOpcodeName(self->instr.uniqueId));
+        return NULL;
+    }
+
+    return PyLong_FromLong(RabbitizerInstruction_getBranchOffsetGeneric(&self->instr));
+}
+
+static PyObject *rabbitizer_type_Instruction_getBranchVramGeneric(PyRabbitizerInstruction *self, UNUSED PyObject *closure) {
+    if (!RabbitizerInstruction_hasOperandAlias(&self->instr, RAB_OPERAND_cpu_label) && !RabbitizerInstruction_hasOperand(&self->instr, RAB_OPERAND_cpu_branch_target_label)) {
+        PyErr_Format(PyExc_RuntimeError, "'%s' instruction does not not perform a branch or a jump.", RabbitizerInstrId_getOpcodeName(self->instr.uniqueId));
+        return NULL;
+    }
+
+    return PyLong_FromUnsignedLong(RabbitizerInstruction_getBranchVramGeneric(&self->instr));
+}
 
 static PyObject *rabbitizer_type_Instruction_getDestinationGpr(PyRabbitizerInstruction *self, UNUSED PyObject *closure) {
     int8_t reg = RabbitizerInstruction_getDestinationGpr(&self->instr);